cleanup and prep work
